Add a kdb dump and a stash file for testing OTP.
This is an imperfect and temporary solution which will be replaced by
something more like what's done for the rest of the tests, that is
building the kdb from scratch.  The OTP plugin is using two special
tl-data types which has no support in the tools though, so an LDAP
backend would be necessary.  Not sure that's the best dependency to
add.